requirements.txt
fastapi==0.111.1
uvicorn==0.23.2
requests==2.32.1
python-dotenv==1.0.1